Teach watchman to ignore buck-out directory
This is useful to ignore a directory that contains only build products
and where file change notifications are unwanted because of the sheer
volume of files.  Not to mention that during tests execution temporary
directories are created and wiped out that may confuse watchman.

This file is provided in the same way as buckversion and must be linked
to the main Buck project directory to have effect, e. g.:

  $ cd <project_dir>
  $ ln -s ../bucklets/watchmanconfig .watchmanconfig

Change-Id: I61ad9f768cce0fec505854008c5f6118f967accc