Add "target_suffix" to gerrit_plugin()
Updating gerrit_plugin.bucklet to match default.defs for Gerrit.

The "target_suffix" parameter is added to the gerrit_plugin()
definition as a way to facilitate a custom "post-build" type of
step, such as with the Gitiles plugin, where the additional static
files need to be moved into the root of the JAR file.

Change-Id: I4e855bdcbb2a2705687e52d3287b7bc8e253391a